Tweak scripts

+++ unity-pp/unity-pp.requestExam.php
@@ -4,12 +4,15 @@ namespace Slothsoft\Devtools\Unity;
 require_once __DIR__ . '/../vendor/autoload.php';
 
 $testatId = 6;
-$testatMessage = "Fix Testat 01-06 (fix sometimes not finding prefab instances)";
+$testatMessage = "Fix Testat 05 (fix mario.GetCurrentJumpSpeed test)";
+$testatDeleteTests = false;
 
 $course = new UnityCourse('repositories.xml', 'results');
 
-foreach ($course->getStudents(true) as $student) {
-    $student->git->reset();
-    $student->unity->deleteFolder('Assets/Tests');
+$course->pullRepositories();
+if ($testatDeleteTests) {
+    foreach ($course->getStudents(true) as $student) {
+        $student->unity->deleteFolder('Assets/Tests');
+    }
 }
 $course->requestTest('tests', $testatId, $testatMessage);
